Slice wcwidth & fix tests
